+++ spec/resources/users_spec.rb
@@ -6,9 +6,15 @@ describe TableauApi::Resources::Users, vcr: { cassette_name: 'users' } do
   describe '#create' do
     # http://onlinehelp.tableau.com/v9.0/api/rest_api/en-us/help.htm#REST/rest_api_ref.htm#Add_User_to_Site%3FTocPath%3DAPI%2520Reference%7C_____7
     it 'can create a user in a site' do
-      user = client.users.create(username: 'test')
+      user = client.users.create(username: 'test', site_role: 'Unlicensed')
       expect(user['id']).to be_a_tableau_id
-      expect(user).to eq('id' => user['id'], 'name' => 'test', 'siteRole' => 'Viewer')
+      expect(user).to eq(
+        'id' => user['id'],
+        'name' => 'test',
+        'siteRole' => 'Unlicensed',
+        'externalAuthUserId' => '',
+        'authSetting' => 'ServerDefault'
+      )
     end
 
     it 'fails with a bad site_role' do
@@ -19,11 +25,12 @@ describe TableauApi::Resources::Users, vcr: { cassette_name: 'users' } do
   describe '#list' do
     # http://onlinehelp.tableau.com/v9.0/api/rest_api/en-us/help.htm#REST/rest_api_ref.htm#Get_Users_on_Site
     it 'can list users in a site' do
+      sleep(15) if VCR.current_cassette.recording?
       user = client.users.list.find do |u|
         u['name'] == 'test'
       end
       expect(user['id']).to be_a_tableau_id
-      expect(user).to eq('id' => user['id'], 'name' => 'test', 'siteRole' => 'Viewer', 'externalAuthUserId' => '')
+      expect(user).to eq('id' => user['id'], 'name' => 'test', 'siteRole' => 'Unlicensed', 'externalAuthUserId' => '')
     end
   end
 
@@ -33,7 +40,7 @@ describe TableauApi::Resources::Users, vcr: { cassette_name: 'users' } do
       user = client.users.list.find do |u|
         u['name'] == 'test'
       end
-      expect(user['siteRole']).to eq('Viewer')
+      expect(user['siteRole']).to eq('Unlicensed')
       user_after_change = client.users.update_user(user_id: user['id'], site_role: 'Publisher')
       expect(user_after_change['siteRole']).to eq('Publisher')
     end